requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
